+++ src/vs/workbench/contrib/issue/electron-sandbox/issueActions.ts
@@ -6,7 +6,7 @@
 import { Action } from 'vs/base/common/actions';
 import * as nls from 'vs/nls';
 import { IssueType } from 'vs/platform/issue/common/issue';
-import { IWorkbenchIssueService } from 'vs/workbench/contrib/issue/electron-sandbox/issue';
+import { IWorkbenchIssueService } from 'vs/workbench/services/issue/common/issue';
 
 export class OpenProcessExplorer extends Action {
 	static readonly ID = 'workbench.action.openProcessExplorer';
